elcuity: abstract out boilerplate for transaction planning
